Update Immutable docs
For an hour or so I was wondering why my form validators and onSubmit function were returning undefined when accessing attributes e.g. `formObj.name`. Turns out, when using the immutable counterparts, things are returned as `Map`. 

Added `formObj.get('name');` so people will get a hint that it works that way.